src/requirements.txt
diff-match-patch==20121119
Django==1.9.7
django-durationfield==0.5.2
django-extensions==1.6.7
django-import-export==0.4.5
django-suit==0.2.19
docopt==0.6.2
jedi==0.9.0
prompt-toolkit==1.0.3
ptpython==0.34
Pygments==2.1.3
six==1.10.0
tablib==0.11.2
wcwidth==0.1.6
Werkzeug==0.11.10
